CACHYOS: tcp_bbr: silence -Wconstant-logical-operand
The `ecn_thresh` and `ecn_factor` params are just consts in a public
version of BBRv3, so evaluating them in conditions as booleans triggers
`-Wconstant-logical-operand` with Clang.

Prepend relevant invocations with double negation to convert values to
booleans explicitly.

The values are always `true`, so these parts of conditions can also be
dropped. Keeping them instead since they are used in an internal version
of BBRv3 to stay as close to the original code as possible.

Link: https://lore.kernel.org/lkml/4616579.LvFx2qVVIh@natalenko.name/T/#u
Link: https://groups.google.com/g/bbr-dev/c/4vwZGw0nxdQ/m/fWt9xmbzAwAJ
Closes: https://codeberg.org/pf-kernel/linux/issues/11
Signed-off-by: Oleksandr Natalenko <oleksandr@natalenko.name>

Link: https://github.com/CachyOS/linux/commit/833ff4f62e754367218f5d6b48e0854e23d006c3